+++ CHANGELOG.md
@@ -1,7 +1,29 @@
+# V5.14.0
+
+## stario
+- Added features
+  * Added MCP31C and TSP650IISK.
+  * Added Auto Interface Select function.
+
+## starioextension
+- Added features
+  * Added MCP31C and TSP650IISK.
+  * Added Auto Interface Select function.
+  * USB HID class (Keyboard mode) support for StarIoExtManager and IPeripheralCommandParser.
+
+## SDK
+- Added Sample Codes
+  * Added sample codes for Auto Interface Select function.
+
+stario (Ver. 2.9.0)  
+starioextension (Ver. 1.15.0)  
+smcloudservices (Ver. 1.4.1)
+
+
 # V5.13.0
 
 Initial release for GitHub
 
-StarIOPort3.1.jar (Ver 2.8.0)  
-starioextension.jar (Ver 1.14.0)  
-smcloudservices.aar (Ver 1.4.1)
+stario (Ver. 2.8.0)  
+starioextension (Ver. 1.14.0)  
+smcloudservices (Ver. 1.4.1)